+++ src/Fluentify.Tests/RecordGeneratorTests/WhenExecuted.cs
@@ -9,6 +9,7 @@ public sealed class WhenExecuted
     [
         typeof(DescriptorAttributeGenerator),
         typeof(FluentifyAttributeGenerator),
+        typeof(HideAttributeGenerator),
         typeof(IgnoreAttributeGenerator),
         typeof(InternalExtensionsGenerator),
         typeof(RecordGenerator),
@@ -27,6 +28,7 @@ public sealed class WhenExecuted
         // Arrange
         Attributes.Descriptor.IsExpectedIn(TestState);
         Attributes.Fluentify.IsExpectedIn(TestState);
+        Attributes.Hide.IsExpectedIn(TestState);
         Attributes.Ignore.IsExpectedIn(TestState);
         Extensions.Internal.IsExpectedIn(TestState);
         Attributes.SkipAutoInitialization.IsExpectedIn(TestState);